lw5/requirements.txt
litestar>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
pydantic[email]
sniffio>=1.3.0